make no-code generic
